bugfix: phoenix tests (#261)
# Description

The default setting (32) for `SALT_BUCKETS` on the Phoenix table` SYSTEM.LOG` is too high and causes regions to get caught in transition. Reduce to a smaller value (currently set to 2 in the test but has been tested up to 8) and set this in the hbase cluster resource. Since we have included the hadoop configuration in `hbase-site.xml` we can point `HADOOP_CONF_DIR` to this, too, as otherwise the wrong location is chosen (by the `sqlline` utility, at least).
`hbase.regionserver.wal.codec` needs to be set to a value to allow phoenix global indices (though this could maybe be a standard setting?)

Jenkins: https://ci.stackable.tech/view/02%20Operator%20Tests%20(custom)/job/hbase-operator-it-custom/61/