Update built to 0.7.1. (#455)
* Update built to 0.7.1.

This requires some code changes as well, as 0.7 contained breaking changes.
This gets rid of GHSA-22q8-ghmq-63vf since the transitive dependency libgit2 is updated as well.

We can also revert some custom code that was used to pull variables from the env at build time, since built was unable to get them from the environment due to the way that build-rust-crate provisioned the environment.

* rustfmt