chore(test): use bash strict mode
According to [kuttl docs], scripts are executed via `sh -c`. The docs do
not suggest if strict mode is enabled, so we should set it. There also
doesn't seem to be a way to choose the shell (whithout calling a shell
youself from within the script).

Without strict mode on, only the final commands exist status is
considered - meaning assertions before the last could fail and never be
picked up.

[kuttl docs]: https://kuttl.dev/docs/testing/steps.html#shell-scripts