chore(test): make assertions more precise with exact matching with grep
We should test exact matches with grep.
By default, `yq` will output strings with double quotes, so those should
be removed with `-r` or `--unwrapScalar` for yq-go (the python yq which
has the same options as jq can use `-r` or `--raw-output`).

Note: I just leart about the `-w` option, which would remove the need
for '^$'.