Adopting according to new readability and code style (#553)
* Adopting according to new readability and code style

* Removing whitespace

* Making shellcheck happy

* chore(test): revert previous styling changes

See: https://github.com/stackabletech/hbase-operator/pull/553#issuecomment-2318215071

* chore(test): wrap variables in double quotes

This would be picked up by `shellcheck` in CI if the script wasn't
inlined. We should consider moving test scripts out of yaml files.

* chore(test): use bash strict mode

According to [kuttl docs], scripts are executed via `sh -c`. The docs do
not suggest if strict mode is enabled, so we should set it. There also
doesn't seem to be a way to choose the shell (whithout calling a shell
youself from within the script).

Without strict mode on, only the final commands exist status is
considered - meaning assertions before the last could fail and never be
picked up.

[kuttl docs]: https://kuttl.dev/docs/testing/steps.html#shell-scripts

* chore(test): add comments to tests

It might seem trivial, but it can help reviewers quickly understand what
the command line is for, and quickly assert whether that lines up with
what the command line is actually doing.

* chore(test): make assertions more precise with exact matching with grep

We should test exact matches with grep.
By default, `yq` will output strings with double quotes, so those should
be removed with `-r` or `--unwrapScalar` for yq-go (the python yq which
has the same options as jq can use `-r` or `--raw-output`).

Note: I just leart about the `-w` option, which would remove the need
for '^$'.

* chore(test): split long commands into multi-line commands, split label options into multiple args

Instead of splitting by line-length, I am:
- Splitting by pipe
- Splitting by flag/option

This is to make reviews simpler when using split panes
(side-by-side).

* chore(test): suggestion: use long-options where possible

I went to town on this one. You could argue that trivial cases (eg: `sed -e`) are ok.
This is just for demonstration.

* chore(test): fix yamllint

* Revert "chore(test): suggestion: use long-options where possible"

This reverts commit fc6fd797c76689c7fd8d2fd115e42fa6e715217e.

* chore(test): fix yamllint

* chore(test): remove -o pipefail bash option

* chore(test): undo ORed labels

---------

Co-authored-by: Nick Larsen <nick.larsen@stackable.tech>